+++ eslint.config.mjs
@@ -1,0 +1,24 @@
+import globals from 'globals';
+import js from '@eslint/js';
+import tseslint from 'typescript-eslint';
+
+export default [
+    js.configs.recommended,
+    ...tseslint.configs.recommended,
+    {
+        languageOptions: {
+            globals: {
+                ...globals.node,
+            },
+        },
+
+        rules: {
+            indent: ['error', 4],
+            'linebreak-style': ['error', 'unix'],
+            quotes: ['error', 'single'],
+            semi: ['error', 'always'],
+            'object-curly-spacing': ['error', 'always'],
+            '@typescript-eslint/no-inferrable-types': 'off',
+        },
+    },
+];